break out Metrics compilation unit
